fix(task): JSON y validaciones
Modificado el JSON para Front y añadidas varias validaciones en el servicio.

+++ src/main/java/com/spring/Service/WorkspaceService.java
@@ -1,17 +1,26 @@
 package com.spring.Service;
 
+import java.lang.reflect.Type;
+import java.util.ArrayList;
 import java.util.Collection;
 
 import javax.transaction.Transactional;
 
+import org.modelmapper.ModelMapper;
+import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.server.ResponseStatusException;
 
-import com.spring.CustomObject.WorkspaceDto;
+import com.spring.CustomObject.ColumnDto;
+import com.spring.CustomObject.TaskForWorkspaceDto;
+import com.spring.CustomObject.UserForWorkspaceDto;
 import com.spring.CustomObject.WorkspaceEditDto;
+import com.spring.CustomObject.WorkspaceWithColumnsDto;
+import com.spring.Model.Column;
 import com.spring.Model.Sprint;
+import com.spring.Model.Task;
 import com.spring.Model.Team;
 import com.spring.Model.User;
 import com.spring.Model.UserAccount;
@@ -32,12 +41,15 @@ public class WorkspaceService extends AbstractService {
 
 	@Autowired
 	private SprintService serviceSprint;
-	
+
 	@Autowired
 	private UserRolService serviceUserRol;
-	
+
 	@Autowired
 	private TeamService serviceTeam;
+
+	@Autowired
+	private TaskService taskService;
 	
 	@Autowired
 	private UserService serviceUser;
@@ -46,9 +58,46 @@ public class WorkspaceService extends AbstractService {
 		return this.repository.findUserRoleByUserAccountAndTeam(userAccount, team);
 	}
 
-	public WorkspaceDto findWorkspaceWithColumns(int id) {
-		Workspace w = this.findOne(id);
-		return new WorkspaceDto(id, w, serviceColumns.findColumnsTasksByWorkspace(id));
+	public WorkspaceWithColumnsDto findWorkspaceWithColumns(int id) {
+		ModelMapper modelMapper = new ModelMapper();
+		Workspace workspace = this.findOne(id);
+		Collection<Task> tasks = this.taskService.findByWorkspace(workspace);
+		
+		Collection<TaskForWorkspaceDto> tasksDtoTodo = new ArrayList<>();
+		Collection<TaskForWorkspaceDto> tasksDtoInProgress = new ArrayList<>();
+		Collection<TaskForWorkspaceDto> tasksDtoDone = new ArrayList<>();
+
+		for (Task task : tasks) {
+			Collection<User> users = task.getUsers();
+			Type collectionUsersDto = new TypeToken<Collection<UserForWorkspaceDto>>() {}.getType();
+			Collection<UserForWorkspaceDto> usersDto = modelMapper.map(users, collectionUsersDto);
+			TaskForWorkspaceDto taskDto = new TaskForWorkspaceDto(task.getId(), task.getTitle(), task.getDescription(), task.getPoints(), usersDto);
+			if(task.getColumn().getName() == "To do") {
+				tasksDtoTodo.add(taskDto);
+			}
+			else if(task.getColumn().getName() == "In progress") {
+				tasksDtoInProgress.add(taskDto);
+			}
+			else {
+				tasksDtoDone.add(taskDto);
+			}
+		}
+		
+		Column columnTodo = this.serviceColumns.findColumnTodoByWorkspace(workspace);
+		ColumnDto columnTodoDto = new ColumnDto(columnTodo.getId(), columnTodo.getName(), tasksDtoTodo);
+		
+		Column columnInProgress = this.serviceColumns.findColumnInprogressByWorkspace(workspace);
+		ColumnDto columnInProgressDto = new ColumnDto(columnInProgress.getId(), columnInProgress.getName(), tasksDtoInProgress);
+
+		Column columnDone = this.serviceColumns.findColumnDoneByWorkspace(workspace);
+		ColumnDto columnDoneDto = new ColumnDto(columnDone.getId(), columnDone.getName(), tasksDtoDone);
+
+		Collection<ColumnDto> columnsDto = new ArrayList<>();
+		
+		columnsDto.add(columnTodoDto);
+		columnsDto.add(columnInProgressDto);
+		columnsDto.add(columnDoneDto);
+		return new WorkspaceWithColumnsDto(workspace.getId(), workspace.getName(), columnsDto);
 	}
 
 	public Collection<Workspace> findWorkspacesByTeam(int team) {
@@ -56,14 +105,14 @@ public class WorkspaceService extends AbstractService {
 	}
 
 	public Workspace findOne(int id) {
-
+		
 		Workspace w = this.repository.findById(id).orElseThrow(
-				() -> new HttpClientErrorException(HttpStatus.BAD_REQUEST, "The workspace requested does not exists"));
+				() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "The workspace requested does not exists"));
 		checkMembers(w.getSprint().getProject().getTeam().getId());
 		return w;
 	}
 
-	public void saveDefaultWorkspace(Sprint sprint) throws Exception {
+	public void saveDefaultWorkspace(Sprint sprint) {
 
 		Workspace workspace = new Workspace();
 		workspace.setName("Default");
@@ -95,27 +144,24 @@ public class WorkspaceService extends AbstractService {
 		return saveTo;
 	}
 
-	public boolean delete(int workspace) {
+	public void delete(int workspace) {
 
-		Boolean check = this.repository.existsById(workspace);
-//
-//		if (check) {
+		boolean check = this.repository.existsById(workspace);
+		if (check) {
 //			checkAuthorityAdmin(workspace);
 //			this.serviceColumns.deleteColumns(workspace);
 			this.repository.deleteById(workspace);
-//		}
-
-		return check;
+		}
 	}
 
 	public void checkMembers(int teamId) {
 		User user = this.serviceUser.getUserByPrincipal();
 		Team team = this.serviceTeam.findOne(teamId);
 		if (user == null) {
-			throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED, "The user must be logged in");
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "The user must be logged in");
 		}
 		if (!this.serviceUserRol.isUserOnTeam(user, team)) {
-			throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED, "The user must belong to the team");
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "The user must belong to the team");
 		}
 
 	}
@@ -125,10 +171,16 @@ public class WorkspaceService extends AbstractService {
 
 		Collection<Team> teams = this.repository.findAllTeamsByUserAccountAdmin(userAccount.getId());
 		Workspace w = this.repository.findById(workspace).orElseThrow(
-				() -> new HttpClientErrorException(HttpStatus.BAD_REQUEST, "The workspace requested does not exists"));
+				() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "The workspace requested does not exists"));
 		if (!teams.contains(w.getSprint().getProject().getTeam())) {
-			throw new HttpClientErrorException(HttpStatus.FORBIDDEN, "You do not own to the team of this workspace.");
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You do not own to the team of this workspace.");
 		}
 	}
+	public void flush() {
+		repository.flush();
+	}
 
+	public Collection<Workspace> findWorkspacesBySprint(int sprint) {
+		return this.repository.findWorkspacesBySprint(sprint);
+	}
 }
